Fix #44 Move all RegexpUtils members to Glob
